refactor: label 로직 추상화
